requirements.txt
fastapi==0.111.0
pydantic==2.7.3
uvicorn==0.27.1
mysql-connector-python
aiofiles
pyjwt
httpx
datetime
python-dotenv
boto3
jinja2==3.1.2
starlette==0.37.2